# src/main/resources/static/public/schema/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "0.0.1-SNAPSHOT",
    "title": "Współogarniacz server REST API",
    "contact": {
      "name": "Paulina Szwed",
      "url": "https://github.com/Delebrith",
      "email": "paulina.szwed.1996@gmail.com"
    }
  },
...
  "securityDefinitions": {
    "JWT": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization"
    }
  },
  "paths": {
...
    "/ranking/{rankingId}/delete":{
      "delete": {
        "operationId": "deleteRanking",
        "tags": [
          "ranking"
        ],
        "summary": "delete ranking",
        "security": [
          {
            "JWT": []
          }
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "rankingId",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64",
            "description": "Task id"
          }
        ],
        "responses": {
          "204": {
            "description": "If ranking was deleted successfully"
          },
          "401": {
            "description": "If user is not authorized to modify this ranking"
          },
          "404": {
            "description": "If ranking is not found"
          },
          "400": {
            "description": "If error occurs"
          }
        }
      }
    },
...
    "/ranking/{rankingId}/find":{
      "get": {
        "operationId": "getRanking",
        "tags": [
          "ranking"
        ],
        "summary": "get ranking",
        "security": [
          {
            "JWT": []
          }
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "rankingId",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64",
            "description": "Ranking id"
          }
        ],
        "responses": {
          "200": {
            "description": "If ranking was modified successfully",
            "schema": {
              "$ref": "#/definitions/Ranking"
            }
          },
          "401": {
            "description": "If user is not authorized to modify this ranking"
          },
          "404": {
            "description": "If ranking is not found"
          },
          "400": {
            "description": "If error occurs"
          }
        }
      }
    },
    "/ranking/{rankingId}/category":{
      "post": {
        "operationId": "addCategory",
        "tags": [
          "ranking"
        ],
        "summary": "Add new category to ranking",
        "security": [
          {
            "JWT": []
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "rankingId",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64",
            "description": "Ranking id"
          },
          {
            "name": "category",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Category"
            },
            "description": "Description of new category"
          }
        ],
        "responses": {
          "201": {
            "description": "If category was created successfully"
          },
          "401": {
            "description": "If user is not authorized to modify this ranking"
          },
          "404": {
            "description": "If ranking is not found"
          },
          "400": {
            "description": "If error occurs"
          }
        }
      }
    },
    "/ranking/{rankingId}/category/{categoryId}":{
      "delete": {
        "operationId": "deleteCategory",
        "tags": [
          "ranking"
        ],
        "summary": "Delete category of ranking",
        "security": [
          {
            "JWT": []
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "categoryId",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64",
            "description": "category id"
          },
          {
            "name": "rankingId",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64",
            "description": "category id"
          }
        ],
        "responses": {
          "204": {
            "description": "If category was deleted successfully"
          },
          "401": {
            "description": "If user is not authorized to modify this category"
          },
          "404": {
            "description": "If category is not found"
          },
          "400": {
            "description": "If error occurs"
          }
        }
      }
    },
    "/ranking/{rankingId}/category/{categoryId}/addPoints":{
      "post": {
        "operationId": "addPoints",
        "tags": [
          "ranking"
        ],
        "summary": "Change user's points in category",
        "security": [
          {
            "JWT": []
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "categoryId",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64",
            "description": "category id"
          },
          {
            "name": "points",
            "in": "query",
            "required": true,
            "type": "integer",
            "format": "int32",
            "minimum": 0,
            "maximum": 100,
            "description": "points"
          },
          {
            "name": "rankingId",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64",
            "description": "category id"
          }
        ],
        "responses": {
          "200": {
            "description": "If score was modified successfully"
          },
          "401": {
            "description": "If user is not authorized to modify this ranking"
          },
          "404": {
            "description": "If category or ranking is not found"
          },
          "400": {
            "description": "If error occurs"
          }
        }
      }
    }
  },
  "definitions": {
...
    "User": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "email": {
          "type": "string",
          "format": "email"
        },
        "password": {
          "type": "string",
          "format": "password"
        },
        "name": {
          "type": "string"
        },
        "authorities": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Authority"
          }
        }
      },
      "required": [
        "email", "name", "password"
      ],
      "example": {
        "email": "user@mail.com",
        "password": "password",
        "name": "User User"
      }
    },
    "Authority": {
      "type": "string",
      "enum": [
        "USER", "ADMIN"
      ]
    },
...
    "Ranking": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "name": {
          "type": "string"
        },
        "categories": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Category"
          }
        }
      },
      "required": [
        "name", "participants"
      ]
    },
    "Category": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "name": {
          "type": "string"
        },
        "scores": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Score"
          }
        }
      },
      "required": [
        "name"
      ]
    },
    "Score": {
      "type": "object",
      "properties": {
        "user": {
          "$ref": "#/definitions/User"
        },
        "points": {
          "type": "integer",
          "format": "int32"
        }
      },
      "required": [
        "user", "points"
      ]
    }
  }
}